+++ config/schedule.rb
@@ -8,3 +8,7 @@ end
 every :friday, at: "4pm" do
   rake "email_digests:weekly"
 end
+
+every 1.day, at: "3pm" do
+  rake "email_digests:create_subscriptions"
+end